py bindings: more generic and versatile setup for installing and testing the python package

+++ src/py_bindings/biguint.rs
@@ -8,7 +8,7 @@ pub struct BigUint {
 #[pymethods]
 impl BigUint {
     #[new]
-    pub fn new(n: u64) -> PyResult<Self> {
+    pub fn new(n: u32) -> PyResult<Self> {
         Ok(Self {
             val: crate::BigUint::from(n),
         })
